storage/flash_area: Great overhaul of flash map API
The commit contains following changes to flash map API:
 - flash_area_open and flash_area_close have been marked as deprecated
   and no longer should be used within Zephyr code;
 - the fa_id, fa_dev_name, fa_device_id have removed from flash_area
   structure;
 - flash device, that the flash area is on, is now linked to the area
   at compile time;
 - functions that are conditionally compiled in with
   CONFIG_FLASH_PAGE_LAYOUT and CONFIG_FLASH_AREA_CHECK_INTEGRITY
   have been moved to dedicated units.
 - adds FLASH_AREA(label) macro that has to be invoked to obtain DTS
   defined area, to be directly used with read/write/erase functions;
   this replaces usage of FLASH_AREA_ID(label) with deprecated
   flash_area_open.
