+++ subsys/storage/flash_map/CMakeLists.txt
@@ -1,6 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
-zephyr_sources(flash_map.c)
+zephyr_sources(flash_map.c flash_map_phase_out.c)
 zephyr_sources_ifndef(CONFIG_FLASH_MAP_CUSTOM flash_map_default.c)
 zephyr_sources_ifdef(CONFIG_FLASH_MAP_SHELL flash_map_shell.c)
+zephyr_sources_ifdef(CONFIG_FLASH_PAGE_LAYOUT flash_map_layout.c)
+zephyr_sources_ifdef(CONFIG_FLASH_AREA_CHECK_INTEGRITY flash_map_integrity.c)
