+++ subsys/storage/flash_map/flash_map_shell.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2019 Intel Corporation
+ * Copyright (c) 2021 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -18,23 +19,26 @@
 LOG_MODULE_REGISTER(flash_map_shell);
 
 extern const struct flash_area *flash_map;
+static int flash_area_id;
 
 static void fa_cb(const struct flash_area *fa, void *user_data)
 {
 	struct shell *shell = user_data;
 
-	shell_print(shell, "%-4d %-8d %-20s  0x%-10x 0x%-12x",
-		    fa->fa_id, fa->fa_device_id, fa->fa_dev_name,
+	shell_print(shell, "%-4d %-17p %-12p %-20s 0x%-8x 0x%-8x",
+		    flash_area_id, fa, fa->fa_dev, fa->fa_dev->name,
 		    fa->fa_off, fa->fa_size);
+	++flash_area_id;
 }
 
 static int cmd_flash_map_list(const struct shell *shell, size_t argc,
 			      char **argv)
 {
-	shell_print(shell, "ID | Device | Device Name"
-		    "       |   Offset   |   Size");
-	shell_print(shell, "-------------------------"
-		    "------------------------------");
+	flash_area_id = 0;
+	shell_print(shell, "ID | Flash area Ptr | Device Ptr | Device Name "
+		    "       |  Offset  | Size");
+	shell_print(shell, "--------------------------------------------------------------------"
+			   "--------");
 	flash_area_foreach(fa_cb, (struct shell *)shell);
 	return 0;
 }
